separate input to db main into a script
